personal_assistant/requirements.txt

langchain==0.0.200
weaviate-client==3.22.1  
asana==2.0.0
requests==2.31.0
beautifulsoup4==4.12.2
PyPDF2==3.0.1
python-docx==0.8.11
pandas==2.0.3
openpyxl==3.1.2
Pillow==10.0.0
python-pptx==0.6.21
pyyaml==6.0.1
pydub==0.25.1
moviepy==1.0.3
ebooklib==0.17.1


pyttsx3==2.90


openai-whisper==20230314


pyaudio==0.2.14
numpy==1.26.0
scipy==1.11.3
sounddevice==0.4.6